Add CORS support and update server port to 5000
Jira Issue ID: WASI-20

Updated the server port to 5000 to resolve the conflict that occured as a result of both
the server and React app running on port 3000.

Enabled CORS in server.js so the backend and frontend can communicate.

Edited comments in server.js to show the correct port.
Removed some unnecessary comments from geoSearch.js function headings.